Add files via upload
Very Basic EDA(even without regression plots, predictive models/basic regressions) and an example of one-hot encoding of columns in a DataFrame. Almost all code is written by me(maybe apart from imported libraries and one of the first cells which was an example). Not a very challenging/interesting lab.